Moved CA loading into the authn config module from cmd
Since this operation also uses env vars for settings, we might as well
do it in the same module and DRY out the main cmd/ module more.
